Make TwobitGitPoller a proper twisted service.
There's a lot of churn here. The important stuff is that the tac
file is no longer necessary. The implementation has been moved into
the gitpoller object that's implemented as a MultiService in the
twobit_gitpoller module. Then we implement the IServiceMaker and
IPlugin in the TwobitGitPoller plugin.

The annoying thing is the way the TimerService doesn't expose the
'now' parameter in the underlying LoopingCall. This means that the
callback is invoked immediately when the TimerService starts. This
sucks because our pollers take a while and twistd won't daemonize
till startService completes.